requirements.txt
netifaces
gpiozero
requests
pillow
numpy
scapy
smbus
spidev
toml
cc1101
websockets
RPi.GPIO
lgpio

# webui stuff
paramiko
flask-socketio
Flask
python-pam
gevent